sockets/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-socketio==5.11.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
